Spawns, Player Labels, Camera Tweaks, Platforms
Screen shakes when a player falls off the edge and the player respawns
on a "respawn platform" and is invincible for a short time. If he moves,
the platform disappears instantly. Also added player labels, one way
colliding platforms and fixed some minor bugs.
Airstunned players regain controls after a defined amount of time, if
they touch the ground they regain it instantly.
It is already possible to add 4 players to the game, only the controls
need to be set up, ID assignment and labeling happens automatically.