Code Cleanup, Asset Reserialization
To make it easier when working with git and Unity, I converted all the
serialization method to "Force Text" instead of binaries. This saves space
and makes for easier merging of conflicts.

I also heavily cleaned up code for PlayerController and Player
identifiers. I renamed StickToPlayer to PlayerIdentifier and made its code
interface easier to use and read.

I also added a SmashGameController to better handle some of the asset
loading. For now it is just managing spawning in player identifiers, but
will be expanded to do much more.